requirements.txt
accelerate==0.31.0
decord==0.6.0
deepspeed==0.16.2
diffusers==0.30.0
einops==0.8.0
huggingface_hub
numpy==1.26.4
omegaconf==2.3.0
onnxruntime-gpu==1.15.1
opencv_python==4.7.0.72
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==10.3.0
safetensors==0.4.3
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.5
transformers==4.46.3
xformers==0.0.26.post1
gradio==5.9.1
moviepy==2.1.1